Merge policy: drop obsolete, 9 years old exception
I believe it's consensual in Tails these days, and has been for years, that "to
be reviewed and get more exposure" is not a strong enough reason to push changes
that break the build of our devel branch. In particular, since this exception
was written, we got ourselves a much simpler and reliable way to build
Tails locally (using Vagrant), and a CI system.