Release process: drop instructions that have a different target audience
This note is meant for the person who will adjust the doc, according to the
output of the previously run code snippet. The RM used to do that work, so it
made sense to have this note in the release process documentation.
Now that our technical writers do this work, this note makes no sense
here. I'm sure our technical writers are able to find issue numbers
in Git without this note.