IUK creation: use a lock file around mksquashfs calls
Let callers like wrap_tails_create_iuks spawn as many tails-create-iuk
jobs as they wish to parallelize the I/Os (rsync), while still ensuring
CPU-intensive tasks (mksquashfs, twice) are run sequentially.

The caller gets to clean up the lock, but doesn't need to create it in
the first place (even if doing so would be preferable, to ensure
ownership and ability to remove it afterwards).