Release process: move ISO history ahead of local IUK builds
There are 2+ GB of images to push through git-annex, which is a huge
bottleneck at the moment. To compensate for that, start the git annex
copy (almost) as soon as the image reproducibility has been verified.

Keep that step after the Torrent file creation step though, as the
latter already moves images (+ their signatures) around a little, paving
the way for the git-annex dance for those who might have $ISOS pointing
to an ISO history checkout. While at it, include some suggested commands
to speed things up.

Once the transfer is started, the release manager can start building
IUKs locally (which might need some adjusting if parallel builds are
desired, until #17657 is acted on).