Fix HTML 5 semantic
https://developer.mozilla.org/en-US/docs/Web/HTML/Element/aside

The HTML <aside> element represents a portion of a document whose
content is only indirectly related to the document's main content.
Asides are frequently presented as sidebars or call-out boxes.

https://developer.mozilla.org/en-US/docs/Web/HTML/Element/nav

The HTML <nav> element represents a section of a page whose purpose is
to provide navigation links, either within the current document or to
other documents. Common examples of navigation sections are menus,
tables of contents, and indexes.