Test suite: ensure we run post_snapshot_restore_hook every time it's needed
I'm seeing scenarios that start with reaching the "tails-greeter" checkpoint
fail: the Greeter is visible on the screen before saving that snapshot, but not
anymore after restoring it.

My current understanding is that's because the workaround introduced in
e41cd4a683b05b1593584baccc4a17792c0b14fb, and later refined by
13cafd5b5fcba33592fec0ccbbdaddeb75054925, was never applied on a snapshot we've
just created, saved, and thus restored (for RAM-only snapshots). It was only
applied when restoring a snapshot that _already existed_ before we started the
current scenario, which is not the case in the situation that exposes the bug
this commit is about.