Test suite: workaround robustness issue
This problem happens when the host system is fast enough to reach
the Greeter before enter_boot_menu_cmdline has noticed that we've
missed the boot menu.

This could plausibly explain some robustness issues we've classified so far as
"missed click in the Greeter" or "GNOME Shell crashed while logging in".