Test suite: call post_vm_start_hook at a time that makes more sense
Previously we would call that function immediately after starting the VM, but
that function does a mouse click, which makes no sense until we have a display
that supports clicking. This became obvious since I did
1881d05af8acb17addebe9af891a2e46c05603db, which now makes post_vm_start_hook
wait for the mouse cursor before exiting, which virtually guarantees we'll miss
the boot menu.

So, instead of calling post_vm_start_hook immediately after starting
the VM, do it after the Greeter is visible and the remote shell is up,
just like we do in post_snapshot_restore_hook already.