Test suite: drop the @fragile tag on the UEFI boot scenario (refs: #13459)
After reading the corresponding issue, my understanding is that
detecting the UEFI firmware screen was causing the problem,
and that code is gone.

Since then a number of other things happened in this department:

 - We've switched from syslinux to GRUB for UEFI boot. This could
   potentially impact the robustness of this test.

 - We've made boot menu and command line handling more robust.

 - UEFI is the common case rather than the exception nowadays,
   so testing it is becoming increasingly more important.

So I think it's worth giving this another try, see what happens,
and be ready to revert this commit.