sync(1) the persistent volume after copying APT data to it (refs: #17278)
I've observed cases when that service started, copied part of the data,
but after the next boot:

 - Not all the data can be found on the persistent volume.
 - The flag file that indicates that this service has finished its job
   is not present.

At this point I don't know whether this service was killed before
it could finish its job, or this service did copy the data but
it was not sync'ed to disk. Let's eliminate the second possibility
and make it easier to reason on this problem.