IUK creation: test if --mksquashfs_lock_file was passed instead of testing its value
I understand we want the code previously guarded by:

   if $self->mksquashfs_lock_file

to be a no-op unless the caller passed the --mksquashfs_lock_file option.

The most accurate way to check that condition is to use the
"has_mksquashfs_lock_file" attribute's predicate that Moo can optionally
generate for us. In most practical cases, what was used before this commit, i.e.
the value of the mksquashfs_lock_file attribute in boolean context, should be
equivalent, but that's due to implicit implementation details rather than thanks
to defined behavior, so let's not take any chances.