Linux kernel upgrade: map source of information to what we're looking for
Thanks to anonym's feedback on #17610, I understood that this mapping
is not obvious, and misunderstandings create confusion, uncertainty,
and make folks do useless extra work.

So let's document better what information one should be looking for *where*,
instead of listing the info we're looking for separately from the sources
of info.