IUK creation: use util-linux' flock(1) to manage the mksquashfs lock
IO::LockedFile is really old-school Perl and seems mostly abandoned upstream,
so it does not seem a great choice for a newly added dependency.

In many other areas of our codebase (tails.git, puppet-tails.git) we've
standardized on the flock prefix command to manage such locks.
It works fine and it's installed by default.