GitLab: use quick action syntax that gives more helpful habits
When one uses the "/label ~xyz" syntax, in the web interface
one gets label completion after typing "~", which is not the case
when one uses the "/label xyz" syntax.

So let's hint contributors towards the former, more usable syntax, even if it
does not matter much here — where it's a matter of copying'n'pasting.