Test suite: also look for locales in /usr/share/locale.
... instead of just /usr/lib/locale, otherwise we won't find the match
"ms:ID" → "ms", so no candidate and the test fails.

Honestly, the difference between these directories isn't entirely
clear to me, but the point here is just to pick a candidate that
works, and with this change the "The Unsafe Browser can be used in all
languages supported in Tails" scenario passes again so I'm happy.