Contributors documentation: add link to Jenkins documentation
That doc is a mix of design notes for sysadmins and usage doc for developers.
Its current location makes sense for the former, but not so much for the latter,
which makes it hard to find for developers: I regularly point developers to this
doc, as they apparently could not find it by themselves.

This should be restructured at some point, but in the meantime,
let's at least have a link from /contribute to that doc.