Security advisory for 4.8: MFSA to be announced.
There is no MFSA out yet, but I really want to release. IIRC our
intention with waiting for the MFSA is to not release a browser based
on Mozilla's code before they do. However, even though there is no
MFSA yet, Firefox 68.10.0esr is available for download, so IMHO we
honor the above intention, so I'll release now, with the MFSA pending.