Test suite: merge scenarios
Moving the only step that the 1st scenario was doing on top of what the 2nd
scenario does, to improve test suite performance. We regress a tiny bit on the
Gherkin expressiveness front but I don't think it's a big deal here: the
"defaults to IMAP" test was mostly valuable when we had code to select IMAP or
POP3 depending on whether the Thunderbird configuration was persistent, which is
not the case anymore.