gdm-error-message: more robust parsing
this takes in consideration the case of revision or other message that
might follow the device code.
I don't have HW that has revisions, so I don't know if they are put
before or after the device code.

The code is now defensive on this: the device code is extracted as the
last block which might look like a device code, and the rest is kept as
device name.