Use simpler term
"Mobile data" is what Android uses:

https://support.google.com/android/answer/9059108?hl=en

GNOME uses "mobile broadband" but we are explicitly referring to this
GUI term in the instructions already.

In this case, I think that it's better to use the simpler and more
common term from the Android terminology than to be consistent with the
GNOME terminology.