Test suite: make the "the Tor Browser has started" step stricter
I've seen a case when the "I start the Tor Browser" step succeeded, but then the
"I open the address $URL" step, run immediately after that, failed with "cannot
find TorBrowserNewTabButton.png on the screen (FindFailed)" a mere second later.

So let's consider "I start the Tor Browser" to be successful only once
the "new tab" button has appeared, which should help ensure the browser
is ready for whatever tests we'll throw at it afterwards.