Translate our Social Contract (#17537)
In the discussion on tails-project, I clarified that the audience for
this document are our users as it states the commitments of our team
towards users.

So it makes sense to translate it *and* move it to /doc until we have a
proper "About Us" section (#17700).

Moving it to /doc also makes it translatable automatically :)