Add basic troubleshooting instructions
We should be more confident now to ask people to re-verify, re-install,
or try a different computer if it doesn't work out-of-the-box.

"Tails fails to start after the Boot Loader" and "the computer stops
responding" overlap but I don't think that it really matters here.
People in this section will be in "brute force" mode and try anything
that seems relevant.