Test suite: improve robustness
I've seen the "Use GNOME Disks to unlock a USB drive that has a basic VeraCrypt
volume with a keyfile" scenario fail with "cannot find
GnomeDisksUnlockDialogKeyfileComboBox.png on the screen (FindFailed)".

The video shows that the previous steps happened super quickly via Dogtail,
while the dialog box was not fully rendered on screen yet, so I understand the
keyfiles combo box was not visible yet when we tried to click on it. Let's wait
a bit for it to appear.

Note that this can go away in Tails 5.0, which will have GNOME Disks ≥ 3.35.2,
which fixes #15952 ⇒ we can use Dogtail for interacting with the keyfiles
combo box too.