Rubocop: target Ruby 2.5 (Buster)
Initially I had set this to 2.3 (Stretch), because most of the code we're
analyzing is meant to be run on our Jenkins isobuilders or isotesters, that
currently run Stretch.

But Rubocop 0.81+ don't support analyzing code targeted at Ruby 2.3,
so let's bump this to 2.5 (Buster), so that at least Rubocop can run.
Rubocop 0.88.0 detects 58 offenses, but that'll be for another day.