Test Suite: bump UnsafeBrowserStartPage.fa.png.
With Tor Browser based on esr78 (refs: #17735) the punctuation
ordering is really wacky on our website's
unsafe_browser_warning.fa.html. The cause seems to be that it has the
language direction (<html dir="...">) set to ltr (which Farsi is)
while displaying text from a rtl language.

For instance, the sentence we are looking for in the image, "The
Unsafe Browser is not anonymous.", is rendered with the dot at the
left edge of the sentence (in general the rule seems to be: if the
right-most character is punctuation, it is moved to the beginning of
the string, e.g. "A.B,C!" → "!A.B,C").

This is the case in both Tor Browser and the Unsafe Browser.

For now, let's bump this image so our tests don't fail.