Fix sorting Intel GPUs last in the "Error starting GDM" message (#17903)
The argument passed to sort_gpus() looks like this:

  [{'Slot': '00:02.0',
    'Class': 'VGA compatible controller [0300]',
    'Vendor-code': '8086',
    'Vendor': 'Intel Corporation',
    'Device-code': '5917',
    'Device': 'UHD Graphics 620',
    'SVendor': 'Lenovo [17aa]',
    'SDevice': 'UHD Graphics 620 [225c]',
    'Rev': '07'}]

So sorting the list based on whether the Vendor ends with "[8086]" did not do
what we want, i.e. sort Intel GPUs last (which we do because the Plymouth
message will be truncated, and we want to display the most relevant GPUs first,
which usually are not the Intel integrated GPU).