Install KeePassXC 2.5.4 from buster-backports (#17286)
Since 2.4.0, KeePassXC uses the LastDir config setting as the default
directory for saving new databases. This allows us to make KeePassXC
databases be stored in /home/amnesia/Persistent by default.

We already set LastDir=/home/amnesia/Persistent in keepassxc.ini, so
installing KeePassXC > 2.4.0 is all we have to do for #17286.