Make code fail hard if it becomes obsolete
I don't think the existence of submodules/tails-workarounds is a good way to
toggle this feature on/off: given how many of us find it hard to deal with
submodules, better avoid adding/removing some too often. E.g. currently we do
have submodules/tails-workarounds but it's empty, which is OK. Let's ensure that
if we ever delete that submodule, we'll drop the corresponding code in
auto/build.