Implement opt-in mechanism for JavaScript version (#17702)
I'm storing the choice in JSON format in the 'custom' variable passed to
the PayPal form. This should work and we will have some confirmation
because it's what I used in the A/B testing for #16830.

But it's bit of a gamble. We won't really know until we get the PayPal
logs from Riseup Labs. Fun!

I tried to pass the email address to PayPal in the 'email' variable but
this only works if you pass ALL variables for the billing form: first
name, phone number, etc. Let's skip doing that for now.