Install trezor packages from buster-backports
Tracking Bullseye here was always meant to be a temporary thing,
until a proper backport was ready. Now there's one :)

We can drop the APT pinning for python3-usb1, since the trezor backport depends
on python3-libusb1 instead, which is in Buster: the package was renamed between
Buster and Bullseye.

refs #17461