Tor Browser: use new trick to avoid mandatory extension signing.
In Tor Browser 10.0a7 the previous method for avoiding the mandatory
extension signature check for HTTPS-Everywhere was replaced with a new
one, which bundles it in omni.ja as a system extension. So we apply
the same trick for uBlock Origin (which lacks a signature because we
install it from Debian).

Details: https://gitlab.torproject.org/tpo/applications/tor-browser/-/issues/40091

Refs: #17933