Revert "Raise the maximum number of loop devices to 32 (refs: #12065)."
This reverts commit 0c048250f33c548313bc7fee25276bd036c7be01, which was about an
experiment that is not running anymore, and that's moot anyway: with #15281, we
won't need more than 2 loop devices to mount SquashFS'es.

This way, we'll get 4 times less udisks "errors" in the logs, when it fails to
determine whether unused loop devices seem to be encrypted or not.

This will mitigate #17351, but it won't fully solve the problem.