Remove useless manual test
Ensuring we keep track of tor releases is FT frontdesk's job.

The best we can hope for, at this stage of the release process, is to realize FT
frontdesk did not do their job, and file an issue to track the missed update.
IMO that's not the sort of things the manual test suite is here for.

AFAIK this manual test has not yielded any value for many years.

refs #10259