Test suite: automate the screen keyboard manual tests
In order to keep the scenario as simple as possible, I did not bother with the
"when activated after the browser / Thunderbird has started" part: this was
meant to be a regression test for #9260, which we fixed 3 years ago,
(2.5 years after it was reported). That bug is not common and easy to
workaround, so I don't think keeping a regression test around for it any longer
is worthwhile.

Closes #10263