Release process: add tip
When the RM lands here, it's easy to despair, especially if their Internet
connection is poor. For example, today I was not thrilled at the idea of
uploading 11 GB of data to an Onion service. But actually it's very fast in the
case at hand, so let's tell the RM there's still hope.