Test suite: handle the fact that the browser address bar is initialized lazily (Persian edition)
The problem we had in English before 391d9eaacd563c79435177e38e63a0dac19b7998
strikes again:

  Initially it says "Search or enter address" and after some actions it says
  "Search with DuckDuckGo or enter address"

This makes the "I open a new tab in the Tor Browser" occasionally fail
in Persian.

In most languages, the pictures in these 2 cases have some common part, to which
we could simply limit our search (sometimes to a somewhat ridiculously small
picture; oh well). But in Persian, there's no common part that I could find:
minor pixel differences are visible even in the part that looks pretty much the
same at first glance. So we need 2 images to support this.

As a side note, the #(exists|find|wait)_any return value is not ideal here:
having to get the match via [1] is a bit too obscure for my taste.