Build system: drop browser localization descriptions for languages not available in the Welcome Screen
We don't install locales for these languages anymore, since users
would not be able to choose them in the Welcome Screen anyway.

Problem is, the "Unsafe Browser works in all supported languages" test suite
step assumes that we ship locales for each of the languages shipped in
the browser localization descriptions file, so it will occasionally fail
with "Could not find a locale for 'ga-IE:IE' (RuntimeError)" or similar.

To fix this, in the browser localization descriptions, only list languages that
are available in the Welcome Screen, and for which we include locales.

refs #17139