Test suite: implement find_any() with real_find().
We recently bumped the timeout in find() from 2 to 5 seconds (in !218
for #17985) which made me notice that it's not very appropriate to
call find() in find_any() and related functions since that timeout
makes us block very long on each image before moving on to the next.

I saw an instance where we use find_any() on five images with a 20
second timeout. This makes it impossible to match the fifth image
since we'll spend 4*5 = 20 seconds on the first four images and thus
timeout before we even try the fifth. Ouch!