check-po-msgstr: Add option to sanitize .po files (tails#17661)
We want to use the `bin/check-po-msgstr` script to run checks before
building the staging website. To make sure the Ikiwiki build will not
break on problematic files, we came up with the solutoin of sanitizing
.po files by removing problematic translations from them.

This commit adds the --sanitize option to such script, which replaces
problematic translations with empty strings.