Revert "Thunderbird: set allow_external_gnupg = true."
This reverts commit 0ff9e084d62a89f52077000046547130df58815c.

We'll document this pref for users that need it (e.g. smartcard users)
instead of having this enabled by default. We want to encourage users
to use Thunderbird's OpenPGP implementation so we don't split them in
two camps with the other using GnuPG, which could result in serious
headache for Help Desk.