Revert "Thunderbird: patch to disable protectHeaders AKA MemoryHole."
This reverts commit 93c54ddc68677983bbbd608b71ec2fdb581f30e7.

So we enable Memory Hole, finally.

Since Thunderbird has this feature enabled by default (and does not
even offer a way to disable it without patching), they are the
pioneers pushing for this, not us, which was part of our original
reluctance to this feature. Also, with a popular email client like
Thunderbird now supporting this out-of-the-box, as well as
k9 (android), mailpile, Evolution, Mail.app, MailMate, Outlook + GnuPG
extension, the problem with unsupported clients seems less of a
problem.