Fix link to the Debian Live manual
We noticed this link was broken 3.5 years ago and since then, we've been hoping
to solve it by self-hosting the version of the Debian Live manual that's
relevant to Tails: tails/sysadmin#12296

Since then, this problem has been reported multiple times: tails/tails!230,
tails/tails#15164, tails/tails#16619, tails/tails#17582. And twice already,
folks have proposed patches that make these links point to the current
version of the Debian Live manual, which sadly is not useful here.

So let's be pragmatic and point to a place that still hosts the version we need:
the Wayback machine.

Closes tails/tails#12261, tails/sysadmin#12296