Build system: ensure we install the required custom packages from our custom APT repo.
We require the packages listed in this hook to be our custom versions,
which we determine by looking whether ".0tails" is present in their
versions. Otherwise we abort the build.

So whenever we bump the APT snapshots and there is a new version
superseding our custom version for some package, we're forced to
either update our custom package, or drop it from the list in this
hook, which makes sense if the superseding version implements the same
fix (or whatever) we required a custom version for in the first place.