Release process: remove now-irrelevant case
We had "and devel" here because previously we were also dealing
with the changelog here, which is not the case anymore.

We never add UDFs on the devel branch (only), so the cleanup
we do in stable, and that will propagate to devel when
we merge stable into devel shortly after, is sufficient.