GitLab CI: check PO files that we did not import from Transifex yet
Goal: spot and report problems there early, instead of letting the RM
discover them when it's too late to fix them.

This is meant to be a temporary workaround. Once we have trusted GitLab CI
runners, instead we should have a job that imports these updated PO files,
commits, and creates a merge request. Then, developers could merge
that merge request as-is if, and only if, it passes the lint-po CI job.

Implementation-wise, none of the GitLab CI -specific YAML features
(include, extends) seems appropriate here, since we effectively need
to merge the `script` array. Hence the use of a YAML anchor.
Other options would be to:

 - KISS and simply duplicate the 2 factorized lines

 - use a shared before_script that can be included by both jobs;
   but so far we've elected to use before_script as a global facility,
   and it would be weird if we overrode it here, and thus had to duplicate
   what our current before_script does

YMMV. There's surely a better way, but what do I know :)