Thunderbird: backport fix for OpenPGP dialog when sending to a recipient without key.
Thunderbird 78.3.3 introduced a regression where a key fetched while
sending email to a recipient that you lack the key for are not listed
so you can "Open details and edit acceptance", requiring to work
around it using the Key Manager.

Reported upstream as:

  https://bugzilla.mozilla.org/show_bug.cgi?id=1675007

but was apparently a duplicate of:

  https://bugzilla.mozilla.org/show_bug.cgi?id=1675285

where I found the patch.