Test suite: wait for GNOME Overview launchers to be ready before interacting with them (fixes: #18016).
I cannot reproduce this locally, but on Jenkins we see the GNOME
Overview opening and sometimes showing the launcher bar to the left in
an incomplete state, with only the top-most icon, Tor Browser, showing
for ~1 second. When the launcher fully loads it changes height,
changing the location of the Tor Browser icon. So there is a race
where we sometimes match the Tor Browser icon and start to click it,
but before we finish the launcher bar readjusts and the icon is moved
away from where we click.