Test suite: bump sleep after selecting language.
Locally I've seen the weird situation where we logged in while Tails
Greeter was in an inconsistent state: we attempted to set the language
as Chinese, but the resulting session only had keyboard and formats
set to Chinese, while the overall locale was still US English. I've
not seen this with other languages, and I do think I notice a higher
latency for Tails Greeter to process update the UI to Chinese which
probably is the cause for this problem. So let's just wait a bit
longer.