Release process: drop dependency on parallel_collect_IUKs
The only reason why we needed parallel_collect_IUKs so far was so that
bin/copy-iuks-to-rsync-server-and-verify, which is run by the RM during the
release process, can fetch every IUK built by Jenkins from a single place.

But parallel_collect_IUKs is painfully slow (tails/sysadmin#17744).

So let's instead teach bin/copy-iuks-to-rsync-server-and-verify how to locate
the IUKs it needs under
https://nightly.tails.boum.org/build_IUKs/configurations/ and download them from
there on rsync.lizard.