Stop installing deb.torproject.org-keyring
The purpose of this package is to keep deb.torproject.org's APT key up-to-date,
which does not work in Tails as we discourage users from updating packages.

On the build/dev side, this package only adds complexity and confusion: we have
to install the APT key ourselves anyway (otherwise live-build would not be able
to install packages we install from deb.torproject.org, before this keyring
package is installed). So this only adds a second place where we need to
maintain the same information, with no benefit. For example, I updated the key
in bfcd3b23e325510c655c35febc9314c7c99413d3, but the deb.torproject.org-keyring
package from the APT snapshot we're using replaces it, in the built images, with
an older, expired one.