Installer: don't make user-facing URLs translatable
We have language negotiation on our website, that ensures the user is served the
localized page if there's one. But

Making these URLs translatable suggests translators they should hard-code the
URL to the localized page (e.g. install/download.fr.html), for which we don't
set up web server rewrite rules when we rename pages. In the past, this has lead
to such links being properly redirected in English, but yielding a 404 in
other languages.