Add tooling for "Don't ask me again" feature for notifications etc.
If a notification want the "Don't ask me again" feature it needs to
settle on a unique identifier, and call `tails-ask-again $identifier`
to learn whether to show it (success) or not (failure). If we pass
`--timeout N` as an argument and the notification was suppressed more
than N days ago, then we reset and allow showing it again. To tell
that a notification should be suppressed, call `tails-ask-again
--dont-ask-again $identifier`.

Closes: #10553