Known issues: link to other potentially affected graphics cards
We keep getting reports about other affected graphics cards in these families,
that are not explicitly listed, and then users did not try the workaround,
because they thought it did not apply to their hardware.

Due to a misunderstanding, so far Help Desk did not forward this info anywhere,
so we did not add those cards to the list. And then other users would report the
same problem over and over again.

So, let's:

 - Give users better means to figure out what workaround they should try,
   so they can solve the problem themselves.

 - As a consequence, decrease the workload of Help Desk (processing such
   reports) and FT (adding cards to the list).