Make *external* TailsData's root user-readable when mounted (closes: #18050)
Since Tails 4.13 when a user has disabled persistence, or plugs
another Tails with a TailsData partition, and tries to mount TailsData
through Nautilus, they are met with an error: 'This location could not
be displayed. You do not have the permissions necessary to view the
contents of "TailsData"'. This is due to #7465 (!210) where we indeed
made the TailsData root world-non-readable, and did not consider this
use case.

Note that this change indeed changes the permission, which might seem
dangerous, but if the Tails is later booted it will self-heal, so all
is ok.