Test suite: delete unused step
The last place we used it for was removed
in 90041e5ed41377484c688be718fdd4821aba824f.

My immediate goal is to make it easier to spot scenarios that are affected
by --old-iso, so I can run only those on my laptop at release time (my local
Jenkins runs the full test suite, but it does not pass --old-iso).