Update roadmap
I now consider this page mainly as a fundraising tool, either to look
for grants or encourage donations.

That's why this year I decided to select less items (9 vs 37) from our
detailed roadmap and focus on the items that I feel are more relevant to
sponsors and donors and that we are the most likely to complete.