Website: disable the htmlscrubber plugin
Previously we made it skip all pages except blueprints.

When I cleaned up the ikiwiki config after removing the blueprints,
I misunderstood the config, commented it out, and as a result
htmlscrubber skips no page at all, which breaks all sorts of stuff.

So let's disable this plugin, which is equivalent to "enabled + skip all pages".

refs tails/tails#18124