onion-grater: workaround tor bug by switch from control socket to control port.
Tor bug: https://gitlab.torproject.org/tpo/core/tor/-/issues/25173

tl;dr: if User is set in torrc (which is the case in Debian and thus
Tails) then the control socket is not opened early enough. The control
port is not affected, so it serves as a workaround for now.