Test suite: adapt to Tor Launcher always starting (refs: #17330)
I picked a rather transparent and implicit approach, by handing the
default Tor Launcher case (direct connection) right in the `Tor is
ready` step. This might be a bit opaque and confusing, however, but
otherwise, in all scenarios/snapshots where we have:

    And the network is plugged
    ...
    Then Tor is ready

we would have to add steps:

    And the network is plugged
    ...
    And the Tor Launcher autostarts
    When I configure a direct connection in Tor Launcher
    Then Tor is ready

which is a bit spammy. I suppose it could be compressed into:

    And the network is plugged
    ...
    Then Tor is ready after I choose a direct connection in Tor Launcher

but I don't like that kind of formulations in Gherkin.