Release process: document best practice to handle release process as a checklist
This is what I've been doing for years and it works well. Let's share this so
the next RM we onboard does not have to reinvent this wheel after N occurrences
of "oops, I did not do step X because I meant to come back to it later and
I forgot".