Unsafe Browser: don't restart tor with restart-tor.
The restart-tor script does not only restart tor, but also monitors
the bootstrap progress for up to 270s, blocking the Unsafe Browser
from exiting (and e.g. be started again). AFAICT there are no benefits
to this monitoring and blocking, just a (admittedly small) drawback,
so let's just restart tor without it.