Test suite: enable garbage collection of failed transient units created by systemd-run
I did not enable this initially because it requires systemd v236 or newer,
and back then we used to support running our test suite on Stretch
(which has v232).

Since #17842 we require Buster or newer to run our test suite,
and Buster has systemd v241, so we can now enable this.

Without --collect, when such a transient unit fails, it would be left behind and
the next test suite run would fail to create a transient unit with the same name
again, until the developer figures out how to clean things up themselves —
which, while entirely feasible, is not exactly trivial.