Ticket gardener: update doc to match current reality
I'm doing this quarterly.

Before we redefined this work, it implied going through _every_ open issue,
hence the lower frequency. Now that it's mostly automated, I find it easier to
do it more often: the situations I deal with are simpler, and the info
I have to share with other folks should arrive at a better time.