Temporarily install the kernel from Bullseye
Linux 5.10.13 was uploaded to buster-backports already, so using that version
matches the spirit of our guidelines. But linux-signed-amd64 has not been built
for the 5.10.13 backport yet, so for now let's install it from Bullseye.

This should be reverted next time we upgrade the kernel, switching
back to the usual buster-backports.

refs #18188