Be more specific and realistic
"Ensure that the workers on your team actually do the work on time"
is not possible, short of the Team manager being ready to do some
of the work themselves, which we can't assume they're able to.

So let's instead describe in more details what a manager *can*,
and probably should, be doing.