Tor Launcher: allow amnesia to start and add suitable .desktop file.
If you close Tor Launcher (perhaps by mistake, which is easily done by
pressing the Esc key) you should be able to relaunch it without having
to reconnect the network, which currently is the only way a user could
force it to restart.

There is no appropriate icon for the .desktop file yet.