Temporarily workaround tor bug by always disabling the seccomp sandbox.
NOTE: This commit should be reverted/dropped before merging this
branch into production!

There is a bug in tor that prevents SAVECONF from working when the
seccomp sandbox is enabled:

  https://gitlab.torproject.org/tpo/core/tor/-/issues/40317

At the moment this breaks pretty much the whole automated test suite:
we successfully create the with-network-logged-in snapshot, but due to
the bug torrc is not updated, so it still has "DisableNetwork" set. So
when we restore the snapshot, and as a consequence restart tor, it
looses the configuration set by Tor Launcher. And since it is not
committed to torrc either they won't be restored, so "DisableNetwork"
is still on and tor won't bootstrap.

More details:

  https://gitlab.tails.boum.org/tails/tails/-/merge_requests/326#note_166438