Test suite: make Dogtail automatically translate strings.
The strings are pulled from the tails domain, so there's a risk that
this breaks testing of non-Tails applications that happen to use the
same English string with a different translation than in the tails
domain. If that happens we'll have to figure out something,
possibly just 'no_translation' flag for the Dogtail methods.